# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AssetMind Backend API",
        "description": "REST API untuk Sistem Pendukung Keputusan Perawatan Aset Industri Perkebunan Kelapa Sawit. Backend Laravel terintegrasi dengan AI Engine (FastAPI) untuk embedding, chatbot RAG, dan rekomendasi pemeliharaan.",
        "contact": {
            "name": "AssetMind",
            "email": "firasyan.daffa123@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (Laravel artisan serve)"
        }
    ],
    "paths": {
...
        "/api/chat/histories": {
            "get": {
                "tags": [
                    "Chatbot"
                ],
                "summary": "Daftar riwayat chat (flat)",
                "description": "Mekanik hanya melihat riwayat miliknya sendiri; teknik melihat semua.",
                "operationId": "chatHistory",
                "responses": {
                    "200": {
                        "description": "Riwayat chat berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
...
        "/api/retrieval-logs": {
            "get": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Daftar retrieval log",
                "description": "Mekanik hanya melihat log miliknya sendiri.",
                "operationId": "retrievalLogIndex",
                "responses": {
                    "200": {
                        "description": "Data retrieval log berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
        "/api/retrieval-logs/{id}": {
            "get": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Detail retrieval log",
                "operationId": "retrievalLogShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail retrieval log berhasil diambil"
                    },
                    "403": {
                        "description": "Bukan milik Anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Hapus retrieval log (khusus teknik)",
                "operationId": "retrievalLogDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieval log berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerSanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "MessageResponse": {
                "properties": {
                    "message": {
                        "description": "Definisi dasar dokumentasi OpenAPI/Swagger untuk AssetMind Backend API.\n\nMemakai PHP 8 Attributes (didukung swagger-php v6).\nAnotasi path tersebar di controller masing-masing.",
                        "type": "string",
                        "example": "Operasi berhasil"
                    }
                },
                "type": "object"
            },
...
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden. Data ini bukan milik Anda."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerSanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum dari endpoint /auth/login (tanpa kata 'Bearer').",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}